feat(ui): UnitDamage, and the third damage-modifier helper
Seven returns: four damage bounds, then the physical damage modifiers.

The modifiers turn out to be the same helpers GetSpellBonusDamage already uses,
asked for school 0 -- physical is just another school, which is why no separate
field exists for it. Two of the three were already ported for that binding; the
third, the percentage modifier at 0x1034, is added beside them. It is a float
where the other two are integers, and it carries the same active-player guard.

Another mistyped pair fixed: minOffhandDamage and maxOffhandDamage were
uint32_t, and the reference takes all four damage bounds through a float load.
No caller had read them yet, so nothing was wrong on screen -- it was waiting to
be. Same class of defect as the attack-power multiplier last cycle, found the
same way, by reading the load rather than the field name.

One asymmetry reproduced rather than smoothed out. A non-player unit reports
0, 0 and 1.0 for the three modifiers -- one, not zero, because the last is a
multiplier. But a player who is NOT the active player takes the helper path
instead, and those helpers answer 0 for anyone but the active player, so the
multiplier comes back 0 there. That is what the reference does; tidying it to
1.0 would have been an invention.

Four more static_asserts on the damage offsets, on the same principle as last
cycle: the offsets are read off the struct, so the build should hold them.

link churn 0 lost, 1 gained. ported 2179->2181, stubs 508->507.
Built clean; not seen running.

+++ src/object/client/CGUnit.hpp
@@ -41,8 +41,11 @@ struct CGUnitData {
     int32_t mountDisplayID;
     float minDamage;
     float maxDamage;
-    uint32_t minOffhandDamage;
-    uint32_t maxOffhandDamage;
+
+    // Floats, like the main-hand pair above. These were uint32_t, which no caller had read yet --
+    // the reference takes all four through a float load.
+    float minOffhandDamage;
+    float maxOffhandDamage;
     int32_t pad2;
     uint32_t petNumber;
     uint32_t petNameTimestamp;
